# openapi.yaml
...
openapi: "3.1.0"
info:
  title: "Filelayer byte-delivery routes"
  version: "0.14.0"
  summary: "The two HTTP routes the Filelayer library serves. There are no others."
  description: "Filelayer is a Node library, not a hosted service. It ships exactly two HTTP route\nhandlers, both in `packages/core/src/delivery.ts`, and this document describes those\ntwo and nothing else.\n\nEverything else in the product — uploading bytes, creating and revoking shares,\nmanaging membership, reading the audit trail — is an in-process method call\n(`fl.files.put()`, `fl.shares.create()`, `fl.orgs.audit()`, …) with no HTTP binding\nthat we ship. You mount those in your own framework, under your own paths, behind\nyour own authentication. If you are looking for `POST /files` in this document, it is\nabsent because it does not exist.\n\nThe two routes below exist as routes rather than as response helpers because both\ncarry a security decision that belongs to the library:\n\n- the response headers are computed from the file record and cannot be turned off, so\n  a developer cannot serve user-uploaded HTML inline from their own origin;\n- the share route reads the password from the request body and refuses, with 400,\n  if any credential appears in the query string — a routing decision, not a response\n  decision, and the reason this is a route.\n\nMount them with `deliveryHandler(fl)`, which is both routes plus a 404 for anything\nelse, and is a complete `node:http` request listener."
...
security: []
...
servers:
...
paths:
  "/f/{fileId}":
    description: "Mounted by `deliveryHandler(fl, { principal })`, whose `filePrefix` option defaults to `/f`. If you mount `fileDownloadRoute(fl, { principal })` directly, its own `prefix` option defaults to `/files`, not `/f` -- pass `prefix: '/f'` to serve the path documented here. Whichever you choose, `publicUrl()` always builds on `/f`, so keep `baseUrl` and the mounted prefix consistent with it."
    get:
      tags:
        - "files"
      operationId: "readFile"
      summary: "Read a file, if the caller is allowed to."
      description: "Re-authorizes on **every** request. That is the whole design: revocation beats a\nlive URL because there is a decision at each request to revoke against.\n\nIdentity is whatever your `principal(req)` callback returns — Filelayer never\nguesses it from a cookie or a header. If you mount the handler without a\n`principal` callback every request is anonymous, which is safe but not a \"public\nmode\": through this route, an anonymous caller reaches only a file carrying an\nexplicit `anonymous` grant. There is no public mode and no `public` column.\n\nThat is true of this route. It is not true of your object store: keys are\n`orgId/fileId` and P2 does not treat them as secrets, so the bucket must be\nprivate or this route is not the only way in.\n\nThis is also the URL returned by `fl.files.put(bytes, { public: true })` and\n`fl.files.publish()`. Calling `fl.files.unpublish()` makes it start returning 404\non the very next request, with no deletion, no key rotation and no cache purge."
      parameters:
        - name: "fileId"
          in: "path"
          required: true
          description: "The opaque file id returned by `put()`. Knowing it grants nothing (P2): it is not an input to any decision."
          schema:
            type: "string"
            format: "uuid"
        - name: "Range"
          in: "header"
          required: false
          description: "A single byte range: `bytes=0-499`, `bytes=500-` or `bytes=-500` (the last 500 bytes). Per RFC 9110 a header this route cannot parse is IGNORED rather than rejected, and the whole object is served under a 200 -- so a malformed value is never an error. Multiple ranges in one header (`bytes=0-9,20-29`) are deliberately ignored for the same reason: answering one of several under a 206 is indistinguishable, from the client side, from an answer to a different question. `If-Range` is not parsed."
          schema:
            type: "string"
          example: "bytes=0-1023"
      responses:
        "200":
          description: "The bytes, streamed, with the full header set. Nothing is buffered in the serving process."
          headers:
            Content-Type:
              description: "The stored content type, after neutralisation. A type that is not a plausible media type, or that contains CR/LF/NUL, is replaced with application/octet-stream. An active type (html, xml, svg, script, xsl, ecmascript) requested inline is also replaced."
              schema:
                type: "string"
            Content-Disposition:
              description: "RFC 6266, with the filename encoded rather than interpolated. `attachment` by default; forced back to `attachment` for an active content type even when `inline` was configured."
              schema:
                type: "string"
            Content-Length:
              description: "Present when the store reported a length for this response. Recomputed from the response rather than trusted from the metadata row."
              schema:
                type: "integer"
            X-Content-Type-Options:
              description: "Always `nosniff`."
              schema:
                type: "string"
            Cache-Control:
              description: "Always `private, no-store, no-cache, must-revalidate, max-age=0`. This is what makes immediate revocation true at the browser and at intermediaries, not only at the origin."
              schema:
                type: "string"
            Pragma:
              description: "Always `no-cache`. The HTTP/1.0 spelling, because corporate proxies are real."
              schema:
                type: "string"
            Expires:
              description: "Always `0`."
              schema:
                type: "string"
            Referrer-Policy:
              description: "Always `no-referrer`. A share link secret is in the URL, so any outbound link inside a delivered document would otherwise leak it in the Referer header."
              schema:
                type: "string"
            Content-Security-Policy:
              description: "Always `default-src 'none'; sandbox`."
              schema:
                type: "string"
            X-Frame-Options:
              description: "Always `DENY`."
              schema:
                type: "string"
            Accept-Ranges:
              description: "Always present on a proxied response. `bytes` normally; `none` when a download cap binds this delivery, because a capped grant has its range dropped and is served whole -- see the 206 description."
              schema:
                type: "string"
                enum:
                  - "bytes"
                  - "none"
            X-Filelayer-Delivery:
              description: "Which delivery mode answered this request: `proxy` or `redirect`."
              schema:
                type: "string"
                enum:
                  - "proxy"
                  - "redirect"
          content:
            "application/octet-stream":
              schema:
                type: "string"
                format: "binary"
        "206":
          description: "The requested byte range. The status is DERIVED from whether the store served a range, never passed in as a parameter -- a partial body under a 200 is undetectable by every HTTP client, so it is not a mistake this API lets a caller make. One case does not reach here: when a download cap binds the grant, the range is dropped and the whole object is returned under a 200 with `Accept-Ranges: none`, because charging a capped grant once per seek would make `maxDownloads: 3` mean \"three seeks\"."
          headers:
            Content-Type:
              description: "The stored content type, after neutralisation. A type that is not a plausible media type, or that contains CR/LF/NUL, is replaced with application/octet-stream. An active type (html, xml, svg, script, xsl, ecmascript) requested inline is also replaced."
              schema:
                type: "string"
            Content-Disposition:
              description: "RFC 6266, with the filename encoded rather than interpolated. `attachment` by default; forced back to `attachment` for an active content type even when `inline` was configured."
              schema:
                type: "string"
            Content-Length:
              description: "Present when the store reported a length for this response. Recomputed from the response rather than trusted from the metadata row."
              schema:
                type: "integer"
            X-Content-Type-Options:
              description: "Always `nosniff`."
              schema:
                type: "string"
            Cache-Control:
              description: "Always `private, no-store, no-cache, must-revalidate, max-age=0`. This is what makes immediate revocation true at the browser and at intermediaries, not only at the origin."
              schema:
                type: "string"
            Pragma:
              description: "Always `no-cache`. The HTTP/1.0 spelling, because corporate proxies are real."
              schema:
                type: "string"
            Expires:
              description: "Always `0`."
              schema:
                type: "string"
            Referrer-Policy:
              description: "Always `no-referrer`. A share link secret is in the URL, so any outbound link inside a delivered document would otherwise leak it in the Referer header."
              schema:
                type: "string"
            Content-Security-Policy:
              description: "Always `default-src 'none'; sandbox`."
              schema:
                type: "string"
            X-Frame-Options:
              description: "Always `DENY`."
              schema:
                type: "string"
            Accept-Ranges:
              description: "Always present on a proxied response. `bytes` normally; `none` when a download cap binds this delivery, because a capped grant has its range dropped and is served whole -- see the 206 description."
              schema:
                type: "string"
                enum:
                  - "bytes"
                  - "none"
            X-Filelayer-Delivery:
              description: "Which delivery mode answered this request: `proxy` or `redirect`."
              schema:
                type: "string"
                enum:
                  - "proxy"
                  - "redirect"
            Content-Range:
              description: "`bytes <start>-<end>/<total>`. The total is the object size as the store reports it, not the `size_bytes` column."
              schema:
                type: "string"
          content:
            "application/octet-stream":
              schema:
                type: "string"
                format: "binary"
        "302":
          description: "Redirect delivery. Only emitted when the instance was configured with `redirectDelivery` AND the verbatim acknowledgement string was supplied; by default only anonymous (published) grants are eligible. Revocation is immediate at decision time plus up to `ttlSeconds` of in-flight window, and `ttlSeconds` is clamped to 300."
          headers:
            Location:
              description: "A short-lived presigned URL at the object store."
              schema:
                type: "string"
                format: "uri"
            Cache-Control:
              description: "For a delivery authorized by an ANONYMOUS grant: `public, max-age=<half the presigned TTL>`. Half, so a cached 302 never hands a client a URL with milliseconds of life left. For every other grant type: `private, no-store, ...`. Nothing that was not already public becomes cacheable by a shared cache."
              schema:
                type: "string"
            Referrer-Policy:
              description: "Always `no-referrer`. The presigned URL is a credential and it is in `Location`."
              schema:
                type: "string"
            X-Content-Type-Options:
              description: "Always `nosniff`."
              schema:
                type: "string"
            X-Filelayer-Delivery:
              description: "Always `redirect`."
              schema:
                type: "string"
                enum:
                  - "redirect"
        "400":
          description: "A credential appeared in the query string. The refused keys, matched case-insensitively, are `password`, `pw`, `pass`, `passwd`, `token`, `secret` and `key`. Refused before any work and without consuming a download."
          headers:
            Content-Type:
              description: "Always `application/json`."
              schema:
                type: "string"
            Cache-Control:
              description: "Always `private, no-store, no-cache, must-revalidate, max-age=0`."
              schema:
                type: "string"
            Pragma:
              description: "Always `no-cache`."
              schema:
                type: "string"
            Referrer-Policy:
              description: "Always `no-referrer`."
              schema:
                type: "string"
            X-Content-Type-Options:
              description: "Always `nosniff`."
              schema:
                type: "string"
          content:
            "application/json":
              schema:
                "$ref": "#/components/schemas/Error"
              example:
                error: "credential_in_query_string"
          x-codes:
            - "credential_in_query_string"
        "404":
          description: "Not found — and deliberately indistinguishable from \"denied\". Internally the exact deny reason is recorded in the audit log; externally every reason that would confirm a file exists collapses to this, so the error surface is not an enumeration oracle across tenants. Covers: no such file, no live grant, a revoked grant, an expired grant, a grant whose download cap is exhausted, an ancestor grant that is no longer live, a forged or unknown share secret, an `as:` naming an identity the project has never seen, a file whose row is deleted, and a file still in `pending` state."
          headers:
            Content-Type:
              description: "Always `application/json`."
              schema:
                type: "string"
            Cache-Control:
              description: "Always `private, no-store, no-cache, must-revalidate, max-age=0`."
              schema:
                type: "string"
            Pragma:
              description: "Always `no-cache`."
              schema:
                type: "string"
            Referrer-Policy:
              description: "Always `no-referrer`."
              schema:
                type: "string"
            X-Content-Type-Options:
              description: "Always `nosniff`."
              schema:
                type: "string"
          content:
            "application/json":
              schema:
                "$ref": "#/components/schemas/Error"
              example:
                error: "not_found"
          x-codes:
            - "not_found"
        "410":
          description: "The file itself has passed its `expiresAt`. Distinct from a dead grant, which is a 404: the file lifecycle gate runs before the grant decision, and a caller who reaches this has already been authorized."
          headers:
            Content-Type:
              description: "Always `application/json`."
              schema:
                type: "string"
            Cache-Control:
              description: "Always `private, no-store, no-cache, must-revalidate, max-age=0`."
              schema:
                type: "string"
            Pragma:
              description: "Always `no-cache`."
              schema:
                type: "string"
            Referrer-Policy:
              description: "Always `no-referrer`."
              schema:
                type: "string"
            X-Content-Type-Options:
              description: "Always `nosniff`."
              schema:
                type: "string"
          content:
            "application/json":
              schema:
                "$ref": "#/components/schemas/Error"
              example:
                error: "gone"
          x-codes:
            - "gone"
        "416":
          description: "A well-formed range that this object cannot satisfy -- a first-pos at or past the end, or a suffix range against a zero-byte object. NOT returned for a malformed range, which is ignored. Before 0.10.0 this case was a 404, which told a caller who was allowed to read the file that it did not exist."
          headers:
            Content-Type:
              description: "Always `application/json`."
              schema:
                type: "string"
            Cache-Control:
              description: "Always `private, no-store, no-cache, must-revalidate, max-age=0`."
              schema:
                type: "string"
            Pragma:
              description: "Always `no-cache`."
              schema:
                type: "string"
            Referrer-Policy:
              description: "Always `no-referrer`."
              schema:
                type: "string"
            X-Content-Type-Options:
              description: "Always `nosniff`."
              schema:
                type: "string"
            Content-Range:
              description: "Always `bytes */<total>`, naming the satisfiable extent. A client that asked past the end did so because it did not know the size; this is the only field that tells it."
              schema:
                type: "string"
          content:
            "application/json":
              schema:
                "$ref": "#/components/schemas/Error"
              example:
                error: "range_not_satisfiable"
          x-codes:
            - "range_not_satisfiable"
        "500":
          description: "An unexpected failure. Never a decision — every authorization outcome is one of the statuses above."
          headers:
            Content-Type:
              description: "Always `application/json`."
              schema:
                type: "string"
            Cache-Control:
              description: "Always `private, no-store, no-cache, must-revalidate, max-age=0`."
              schema:
                type: "string"
            Pragma:
              description: "Always `no-cache`."
              schema:
                type: "string"
            Referrer-Policy:
              description: "Always `no-referrer`."
              schema:
                type: "string"
            X-Content-Type-Options:
              description: "Always `nosniff`."
              schema:
                type: "string"
          content:
            "application/json":
              schema:
                "$ref": "#/components/schemas/Error"
              example:
                error: "internal"
          x-codes:
            - "internal"
...
components:
  schemas:
    Error:
      type: "object"
      required:
        - "error"
      description: "Every error on these routes has this shape. `error` is a stable machine-readable code. The internal reason for a denial is deliberately NOT included — it is written to the audit log, where the operator who needs it can see it, and withheld from the caller, who would otherwise be able to distinguish \"does not exist\" from \"you may not have it\"."
      properties:
        error:
          type: "string"
          description: "The stable error code."
          enum:
            - "not_found"
            - "gone"
            - "password_required"
            - "credential_in_query"
            - "credential_in_query_string"
            - "bad_request"
            - "payload_too_large"
            - "internal"
        retry:
          type: "object"
          description: "Present only on 401. How to re-issue the request successfully."
          properties:
            method:
              type: "string"
              enum:
                - "POST"
            field:
              type: "string"
              enum:
                - "password"